Add BusObject interface.
This makes the Object bits easier to mock or fake. Also extracted
the Object bits from the Call bits to make the code easier to
navigate.

This resolves #25.
